+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
++ Added `Trace.logLine` extension method
+*Breaking changes:*
++ Removed getLineInfo function
+
 ## 1.0.0
 
 + `ConsoleBacked` and `FileBackend`